+++ lib/src/data/converter/message/message_hive_extention_converter.dart
@@ -1,7 +1,6 @@
 import 'package:amity_sdk/src/core/enum/amity_message_type.dart';
 import 'package:amity_sdk/src/data/data_source/local/hive_entity/message_hive_entity_18.dart';
 import 'package:amity_sdk/src/domain/domain.dart';
-import 'package:amity_sdk/src/domain/model/message/amity_message.dart';
 
 /// [MessageHiveExtensionConverter]
 extension MessageHiveExtensionConverter on MessageHiveEntity {