+++ lib/src/domain/usecase/poll/create_poll_usecase.dart
@@ -1,6 +1,5 @@
 import 'package:amity_sdk/src/core/core.dart';
 import 'package:amity_sdk/src/domain/domain.dart';
-import 'package:amity_sdk/src/domain/repo/poll_repo.dart';
 
 /// Get Poll Use Case
 class CreatePollUseCase extends UseCase<AmityPoll, CreatePollRequest> {